Move cluster_single_parameters from cubids to utils.
